requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
passlib==1.7.4
PyMySQL==1.0.2
requests==2.26.0
SQLAlchemy==1.4.25
sqlalchemy-datatables==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
oauthlib===3.1.1
pyOpenSSL==21.0.0
flask_login==0.5.0